Create Count_Text_ Excel
This code helps in reading an excel file with a column full of comments and then generates an excel file that shares the word and the frequency of the word. 

Application:
Understanding the customer emotions through their comments: This can be used to analyze the comments of various users and can help in understanding which word was repeated the most.